+++ src/client/app/components/+about/about-apple-pear.component.ts
@@ -2,7 +2,7 @@
 import { Component } from '@angular/core';
 
 // framework
-import { BaseComponent } from '../../framework/core/src/base.component';
+import { BaseComponent } from '../../framework/core/core.module';
 
 // app
 import { routeAnimation } from '../../app.animations';
